+++ src/main/java/be/ugent/rml/MappingFactory.java
@@ -14,7 +14,6 @@ import be.ugent.rml.termgenerator.TermGenerator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.naming.Name;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -31,6 +30,9 @@ public class MappingFactory {
     private Term triplesMap;
     private QuadStore store;
     private List<PredicateObjectGraphMapping> predicateObjectGraphMappings;
+    // This boolean is true when the double in a reference need to be ignored.
+    // For example, when accessing data in a RDB.
+    private boolean ignoreDoubleQuotes;
     protected Logger logger = LoggerFactory.getLogger(this.getClass());
 
     public MappingFactory(FunctionLoader functionLoader) {
@@ -43,6 +45,7 @@ public class MappingFactory {
         this.subjectMappingInfo = null;
         this.predicateObjectGraphMappings = new ArrayList<>();
         this.graphMappingInfos = null;
+        this.ignoreDoubleQuotes = this.areDoubleQuotesIgnored(store, triplesMap);
 
         parseSubjectMap();
         parsePredicateObjectMaps();
@@ -71,7 +74,7 @@ public class MappingFactory {
 
                     //checking if we are dealing with a Blank Node as subject
                     if (!termTypes.isEmpty() && termTypes.get(0).equals(new NamedNode(NAMESPACES.RR  + "BlankNode"))) {
-                        SingleRecordFunctionExecutor executor = RecordFunctionExecutorFactory.generate(store, subjectmap, true);
+                        SingleRecordFunctionExecutor executor = RecordFunctionExecutorFactory.generate(store, subjectmap, true, ignoreDoubleQuotes);
 
                         if (executor != null) {
                             generator = new BlankNodeGenerator(executor);
@@ -80,7 +83,7 @@ public class MappingFactory {
                         }
                     } else {
                         //we are not dealing with a Blank Node, so we create the template
-                        generator = new NamedNodeGenerator(RecordFunctionExecutorFactory.generate(store, subjectmap, true));
+                        generator = new NamedNodeGenerator(RecordFunctionExecutorFactory.generate(store, subjectmap, true, ignoreDoubleQuotes));
                     }
                 } else {
                     SingleRecordFunctionExecutor functionExecutor = parseFunctionTermMap(functionValues.get(0));
@@ -186,9 +189,9 @@ public class MappingFactory {
 
         if (functionValues.isEmpty()) {
             boolean encodeIRI = termType != null && termType.getValue().equals(NAMESPACES.RR + "IRI");
-            SingleRecordFunctionExecutor executor = RecordFunctionExecutorFactory.generate(store, objectmap, encodeIRI);
+            SingleRecordFunctionExecutor executor = RecordFunctionExecutorFactory.generate(store, objectmap, encodeIRI, ignoreDoubleQuotes);
 
-            if (executor != null) {
+            if (parentTriplesMaps.isEmpty() && parentTermMaps.isEmpty()) {
                 TermGenerator oGen;
 
                 if (termType.equals(new NamedNode(NAMESPACES.RR + "Literal"))) {
@@ -204,7 +207,12 @@ public class MappingFactory {
                 } else if (termType.equals(new NamedNode(NAMESPACES.RR + "IRI"))) {
                     oGen = new NamedNodeGenerator(executor);
                 } else {
-                    oGen = new BlankNodeGenerator(executor);
+                    if (executor == null) {
+                        // This will generate Blank Node with random identifiers.
+                        oGen = new BlankNodeGenerator();
+                    } else {
+                        oGen = new BlankNodeGenerator(executor);
+                    }
                 }
 
                 objectMapCallback.accept(new MappingInfo(objectmap, oGen), "child");
@@ -232,11 +240,12 @@ public class MappingFactory {
                         FunctionModel equal = functionLoader.getFunction(new NamedNode("http://example.com/idlab/function/equal"));
                         Map<String, Object[]> parameters = new HashMap<>();
 
-                        SingleRecordFunctionExecutor parent = new ReferenceExtractor(parents.get(0));
+                        boolean ignoreDoubleQuotesInParent = this.areDoubleQuotesIgnored(store, parentTriplesMap);
+                        SingleRecordFunctionExecutor parent = new ReferenceExtractor(parents.get(0), ignoreDoubleQuotesInParent);
                         Object[] detailsParent = {"parent", parent};
                         parameters.put("http://users.ugent.be/~bjdmeest/function/grel.ttl#valueParameter", detailsParent);
 
-                        SingleRecordFunctionExecutor child = new ReferenceExtractor(childs.get(0));
+                        SingleRecordFunctionExecutor child = new ReferenceExtractor(childs.get(0), ignoreDoubleQuotes);
                         Object[] detailsChild = {"child", child};
                         parameters.put("http://users.ugent.be/~bjdmeest/function/grel.ttl#valueParameter2", detailsChild);
 
@@ -301,7 +310,7 @@ public class MappingFactory {
             TermGenerator generator;
 
             if (functionValues.isEmpty()) {
-                SingleRecordFunctionExecutor executor = RecordFunctionExecutorFactory.generate(store, graphMap, true);
+                SingleRecordFunctionExecutor executor = RecordFunctionExecutorFactory.generate(store, graphMap, true, ignoreDoubleQuotes);
 
                 if (termType == null || termType.equals(new NamedNode(NAMESPACES.RR + "IRI"))) {
                     generator = new NamedNodeGenerator(executor);
@@ -346,7 +355,7 @@ public class MappingFactory {
 
             if (functionValues.isEmpty()) {
                 predicateMappingInfos.add(new MappingInfo(predicateMap,
-                        new NamedNodeGenerator(RecordFunctionExecutorFactory.generate(store, predicateMap, false))));
+                        new NamedNodeGenerator(RecordFunctionExecutorFactory.generate(store, predicateMap, false, ignoreDoubleQuotes))));
             } else {
                 SingleRecordFunctionExecutor functionExecutor = parseFunctionTermMap(functionValues.get(0));
 
@@ -449,7 +458,7 @@ public class MappingFactory {
             List<Term> functionValues = Utils.getObjectsFromQuads(store.getQuads(languageMap, new NamedNode(NAMESPACES.FNML + "functionValue"), null));
 
             if (functionValues.isEmpty()) {
-                executors.add(RecordFunctionExecutorFactory.generate(store, languageMap, false));
+                executors.add(RecordFunctionExecutorFactory.generate(store, languageMap, false, ignoreDoubleQuotes));
             } else {
                 executors.add(parseFunctionTermMap(functionValues.get(0)));
             }
@@ -516,4 +525,31 @@ public class MappingFactory {
         return list;
     }
 
+    /**
+     * This function returns true if double quotes should be ignored in references.
+     * @param store The store with the RML rules.
+     * @param triplesMap The Triples Map that should be checked.
+     * @return true if double quotes should be ignored in references, else false.
+     */
+    private boolean areDoubleQuotesIgnored(QuadStore store, Term triplesMap) {
+        List<Term> logicalSources = Utils.getObjectsFromQuads(store.getQuads(triplesMap, new NamedNode(NAMESPACES.RML + "logicalSource"), null));
+
+        if (!logicalSources.isEmpty()) {
+            Term logicalSource = logicalSources.get(0);
+
+            List<Term> sources = Utils.getObjectsFromQuads(store.getQuads(logicalSource, new NamedNode(NAMESPACES.RML + "source"), null));
+
+            if (!sources.isEmpty()) {
+                Term source = sources.get(0);
+
+                if (! (sources.get(0) instanceof Literal)) {
+                    List<Term> sourceType = Utils.getObjectsFromQuads(store.getQuads(source, new NamedNode(NAMESPACES.RDF + "type"), null));
+
+                    return sourceType.get(0).getValue().equals(NAMESPACES.D2RQ + "Database");
+                }
+            }
+        }
+
+        return false;
+    }
 }